Move or delete the remaining helper files in src
