--- packages/yoastseo/spec/helpers/getLanguagesWithWordFormSupportSpec.js
@@ -1,7 +1,0 @@
-import getLanguagesWithWordFormSupport from "../../src/helpers/_todo/getLanguagesWithWordFormSupport";
-
-describe( "Checks which languages have morphology support in YoastSEO.js", function() {
-	it( "returns an array of languages that have morphology support", function() {
-		expect( getLanguagesWithWordFormSupport() ).toEqual( [ "en", "de", "nl", "es", "fr", "ru", "it", "pt", "id", "pl", "ar" ] );
-	} );
-} );
