+++ packages/yoastseo/src/config/transliterations.js
@@ -1,4 +1,4 @@
-import getLanguage from '../helpers/_todo/getLanguage.js';
+import getLanguage from '../languageProcessing/helpers/language/getLanguage.js';
 import { isUndefined } from "lodash-es";
