--- packages/yoastseo/src/helpers/_todo/index.js
@@ -1,11 +1,0 @@
-import scoreToRating from "../../scoring/interpreters/scoreToRating";
-import { measureTextWidth } from "../createMeasurementElement";
-import getLanguagesWithWordFormSupport from "./getLanguagesWithWordFormSupport";
-
-export {
-	measureTextWidth,
-	scoreToRating,
-	getLanguagesWithWordFormSupport,
-};
-
-export * from "../inRange";
